Implement generateReport end-point and make necessary changes

+++ src/main/java/com/example/RetailService/entity/Transaction.java
@@ -6,6 +6,7 @@ import lombok.*;
 import org.springframework.data.annotation.Id;
 import org.springframework.data.mongodb.core.mapping.Document;
 
+import java.math.BigDecimal;
 import java.util.Date;
 
 @Data
@@ -20,7 +21,7 @@ public class Transaction {
     String id;
     Product[] products;
     TransactionType type;
-    Double total;
+    BigDecimal total;
     Date date;
     String userId;
 }
